Refactor interval_collision helpers
