+++ include/physics/collider.h
@@ -9,18 +9,18 @@ typedef enum {
   COLLIDER_PLANE,
 } ColliderType;
 
-struct Collider {
-  COLLIDER_TYPE type;
-  // Thanks Low Level
-  union {
-    struct AABB aabb;
-    struct Plane plane;
-  } data;
-  vec3 position;
-  // Spheres wouldn't care about rotation... but whatever
-  vec3 rotation;
-  vec3 velocity;
-};
+// struct Collider {
+//   COLLIDER_TYPE type;
+//   // Thanks Low Level
+//   union {
+//     struct AABB aabb;
+//     struct Plane plane;
+//   } data;
+//   vec3 position;
+//   // Spheres wouldn't care about rotation... but whatever
+//   vec3 rotation;
+//   vec3 velocity;
+// };
 
 // struct PlaneCollider {
 //   vec3 normal;
